requirements/test.txt
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
tox==4.15.1
model_bakery==1.18.1
django-stubs==5.0.2